requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.41
botocore==1.29.41
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
Django==4.1.3
django-celery-results==2.4.0
django-db==0.0.7
django-extensions==3.2.1
django-extras==0.3
django-filter==22.1
django-phone-field==1.8.1
django-rest-swagger==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-dynamic-fields==0.4.0
Faker==15.3.4
graphviz==0.20.1
idna==3.4
importlib-metadata==5.1.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
Naked==0.1.32
numpy==1.23.5
openapi-codec==1.3.2
pandas==1.5.2
prompt-toolkit==3.0.36
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.16.0
pydotplus==2.0.2
PyJWT==2.6.0
psycopg2-binary
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
redis==4.4.0
requests==2.28.1
s3transfer==0.6.0
secure-smtplib==0.1.1
shellescape==3.8.1
simplejson==3.18.0
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
zipp==3.11.0
django-filter
djangorestframework
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-db==0.0.7
django-extensions==3.2.1
django-extras==0.3
django-filter==22.1
django-phone-field==1.8.1
django-rest-swagger==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-dynamic-fields==0.4.0
python-decouple